requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.1.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.2.2=pypi_0
accelerate=1.8.1=pypi_0
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.12.13=pypi_0
aiosignal=1.3.2=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
aom=3.6.0=h6a678d5_0
async-timeout=5.0.1=pypi_0
attrs=25.3.0=pypi_0
augmentor=0.2.12=pypi_0
blas=1.0=mkl
blosc=1.21.3=h6a678d5_0
brotli=1.0.9=h5eee18b_9
brotli-bin=1.0.9=h5eee18b_9
brotli-python=1.0.9=py39h6a678d5_9
brunsli=0.1=h2531618_0
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2025.2.25=h06a4308_0
certifi=2025.4.26=py39h06a4308_0
cfitsio=3.470=h5893167_7
charls=2.2.0=h2531618_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.8=pypi_0
coloredlogs=15.0.1=pypi_0
contextlib2=21.6.0=pypi_0
contourpy=1.2.1=py39hdb19cb5_1
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cuda-version=12.8=3
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h52b45da_1
dav1d=1.2.1=h5eee18b_0
docker-pycreds=0.4.0=pypi_0
dominate=2.9.1=pypi_0
einops=0.8.1=pypi_0
ema-pytorch=0.7.7=pypi_0
et-xmlfile=2.0.0=pypi_0
eval-type-backport=0.2.2=pypi_0
exceptiongroup=1.2.2=pypi_0
expat=2.6.4=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py39h06a4308_0
flake8=7.1.2=pypi_0
flatbuffers=25.2.10=pypi_0
fontconfig=2.14.1=h55d465d_3
fonttools=4.55.3=py39h5eee18b_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.7.0=pypi_0
fsspec=2025.5.1=pypi_0
giflib=5.2.2=h5eee18b_0
gitdb=4.0.12=pypi_0
gitpython=3.1.44=pypi_0
gmp=6.3.0=h6a678d5_0
gmpy2=2.2.1=py39h5eee18b_0
gnutls=3.6.15=he1e5248_0
grpcio=1.62.2=py39h6a678d5_0
h5py=3.12.1=py39h5842655_1
hdf5=1.14.5=h2b7332f_2
hf-xet=1.1.5=pypi_0
huggingface-hub=0.33.0=pypi_0
humanfriendly=10.0=pypi_0
icu=73.1=h6a678d5_0
idna=3.7=py39h06a4308_0
imagecodecs=2024.9.22=py39h73404f3_0
imageio=2.37.0=py39h06a4308_0
importlib-metadata=8.5.0=py39h06a4308_0
importlib_resources=6.4.0=py39h06a4308_0
iniconfig=2.0.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.5=py39h06a4308_0
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_3
jxrlib=1.1=h7b6447c_2
kiwisolver=1.4.4=py39h6a678d5_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lazy_loader=0.4=py39h06a4308_0
lcms2=2.16=hb9589c4_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=4.0.0=h6a678d5_0
libabseil=20240116.2=cxx17_h6a678d5_0
libaec=1.1.3=h6a678d5_0
libavif=1.1.1=h5eee18b_0
libbrotlicommon=1.0.9=h5eee18b_9
libbrotlidec=1.0.9=h5eee18b_9
libbrotlienc=1.0.9=h5eee18b_9
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.13.0.11=0
libcups=2.4.2=h2d74bed_1
libcurand=10.3.9.55=0
libcurl=8.12.1=hc9e6f67_0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.22=h5eee18b_0
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libglib=2.78.4=hdc74915_0
libgomp=11.2.0=h1234567_1
libgrpc=1.62.2=h2d74bed_0
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnghttp2=1.57.0=h2d74bed_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libpq=17.2=hdbd6064_0
libprotobuf=4.25.3=he621ea3_0
libssh2=1.11.1=h251f7ec_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=hffd6297_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_1
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h097e994_2
libxml2=2.13.5=hfdd30dd_0
libzopfli=1.0.3=he6710b0_0
lightning-utilities=0.14.3=pypi_0
llvm-openmp=14.0.6=h9e868ea_0
lmdb=1.6.2=pypi_0
lxml=5.3.1=pypi_0
lz4-c=1.9.4=h6a678d5_1
markdown=3.8=py39h06a4308_0
markupsafe=3.0.2=py39h5eee18b_0
matplotlib=3.9.2=py39h06a4308_1
matplotlib-base=3.9.2=py39hbfdbfaf_1
mccabe=0.7.0=pypi_0
medpy=0.5.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_2
mkl_fft=1.3.11=py39h5eee18b_0
mkl_random=1.2.8=py39h1128e8f_0
ml-collections=0.1.1=pypi_0
mpc=1.3.1=h5eee18b_0
mpfr=4.2.1=h5eee18b_0
mpmath=1.3.0=py39h06a4308_0
multidict=6.5.0=pypi_0
mypy=1.15.0=pypi_0
mypy-extensions=1.0.0=pypi_0
mysql=8.4.0=h29a9f33_1
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.2.1=py39h06a4308_0
ninja=1.11.1.4=pypi_0
numpy=1.26.4=py39h5f9d8c6_0
numpy-base=1.26.4=py39hb5e798b_0
omegaconf=2.3.0=pypi_0
onnx=1.17.0=pypi_0
onnxruntime=1.19.2=pypi_0
opencv-python-headless=4.11.0.86=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openldap=2.6.4=h42fbc30_0
openpyxl=3.1.5=pypi_0
openssl=3.0.16=h5eee18b_0
packaging=24.2=py39h06a4308_0
pandas=2.2.3=pypi_0
pcre2=10.42=hebb0a14_1
pillow=11.1.0=py39hcea889d_0
pip=25.0=py39h06a4308_0
piqa=1.3.2=pypi_0
platformdirs=4.3.7=pypi_0
pluggy=1.5.0=pypi_0
propcache=0.3.2=pypi_0
protobuf=6.30.2=pypi_0
psutil=7.0.0=pypi_0
pybind11-abi=4=hd3eb1b0_1
pycocotools=2.0.8=pypi_0
pycodestyle=2.12.1=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pydicom=2.4.4=pypi_0
pyflakes=3.2.0=pypi_0
pyparsing=3.2.0=py39h06a4308_0
pyqt=6.7.1=py39h6a678d5_0
pyqt6-sip=13.9.1=py39h5eee18b_0
pysocks=1.7.1=py39h06a4308_0
pytest=8.3.5=pypi_0
python=3.9.21=he870216_1
python-dateutil=2.9.0post0=py39h06a4308_2
python-docx=1.1.2=pypi_0
pytorch=2.4.0=py3.9_cuda11.8_cudnn9.1.0_0
pytorch-cuda=11.8=h7e8668a_6
pytorch-fid=0.3.0=pypi_0
pytorch-lightning=2.5.1.post0=pypi_0
pytorch-mutex=1.0=cuda
pytz=2025.2=pypi_0
pywavelets=1.6.0=pypi_0
pyyaml=6.0.2=pypi_0
qtbase=6.7.2=hdaa5aa8_1
qtdeclarative=6.7.2=h6a678d5_0
qtsvg=6.7.2=he621ea3_0
qttools=6.7.2=h80c7b02_0
qtwebchannel=6.7.2=h6a678d5_0
qtwebsockets=6.7.2=h6a678d5_0
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
requests=2.32.3=py39h06a4308_1
safetensors=0.5.3=pypi_0
scikit-image=0.24.0=py39h1128e8f_0
scikit-learn=1.6.1=pypi_0
scipy=1.13.1=py39h5f9d8c6_1
segment-anything=1.0=dev_0
sentry-sdk=2.27.0=pypi_0
setproctitle=1.3.6=pypi_0
setuptools=75.8.0=py39h06a4308_0
simpleitk=2.5.0=pypi_0
sip=6.10.0=py39h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.2=pypi_0
snappy=1.2.1=h6a678d5_0
sqlite=3.45.3=h5eee18b_0
sympy=1.13.3=py39h06a4308_1
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.19.0=py39h06a4308_0
tensorboard-data-server=0.7.0=py39h52d8a92_1
tensorboardx=2.6.2.2=pypi_0
threadpoolctl=3.6.0=pypi_0
tifffile=2023.4.12=py39h06a4308_0
tk=8.6.14=h39e8969_0
tomli=2.0.1=py39h06a4308_0
torchaudio=2.4.0=py39_cu118
torchmetrics=1.7.3=pypi_0
torchsummary=1.5.1=pypi_0
torchtriton=3.0.0=py39
torchvision=0.19.0=py39_cu118
tornado=6.4.2=py39h5eee18b_0
tqdm=4.67.1=py39h2f386ee_0
typing_extensions=4.12.2=py39h06a4308_0
tzdata=2025.2=pypi_0
unicodedata2=15.1.0=py39h5eee18b_1
unzip=1.0.0=pypi_0
urllib3=2.3.0=py39h06a4308_0
wandb=0.19.10=pypi_0
werkzeug=3.1.3=py39h06a4308_0
wheel=0.45.1=py39h06a4308_0
xcb-util-cursor=0.1.4=h5eee18b_0
xz=5.6.4=h5eee18b_1
yarl=1.20.1=pypi_0
zfp=1.0.0=h6a678d5_0
zipp=3.21.0=py39h06a4308_0
zlib=1.2.13=h5eee18b_1
zlib-ng=2.0.7=h5eee18b_0
zstd=1.5.6=hc292b87_0
